services/auth/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
dotenv-python==0.0.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.41
email-validator==2.2.0
zxcvbn==4.5.0
psycopg2-binary==2.9.10
pydantic-settings==2.9.1
alembic==1.16.1
python-jose[cryptography]==3.5.0